Fixed issue with the main element.
WHen I introduced the main element, I put it in such a place that it
overlapped with another element.  The problem is that the id=mayura
div intersects (instead of subsumes) the main element.

I fixed it by duplicating the css that was for id=mayura into that for
the element main.  And then ended the id=mayura before the main
element starts.  Whatta hack.